[BAEL-16531] - Move all the existing articles on "optional" under java-optional
