+++ core-java-modules/core-java-9-improvements/README.md
@@ -5,7 +5,6 @@ This module contains articles about the improvements to core Java features intro
 ### Relevant Articles:
 
 - [New Stream Collectors in Java 9](http://www.baeldung.com/java9-stream-collectors)
-- [Java 9 Optional API Additions](https://www.baeldung.com/java-9-optional)
 - [Java 9 Convenience Factory Methods for Collections](https://www.baeldung.com/java-9-collections-factory-methods)
 - [Java 9 Stream API Improvements](https://www.baeldung.com/java-9-stream-api)
 - [Java 9 java.util.Objects Additions](https://www.baeldung.com/java-9-objects-new)
