+++ core-java-modules/core-java-9/README.md
@@ -4,13 +4,11 @@ This module contains articles about Java 9 core features
 
 ### Relevant Articles:
 
-- [Method Handles in Java](http://www.baeldung.com/java-method-handles)
-- [Introduction to Chronicle Queue](http://www.baeldung.com/java-chronicle-queue)
-- [Optional orElse Optional](http://www.baeldung.com/java-optional-or-else-optional)
+- [Method Handles in Java](https://www.baeldung.com/java-method-handles)
+- [Introduction to Chronicle Queue](https://www.baeldung.com/java-chronicle-queue)
 - [Iterate Through a Range of Dates in Java](https://www.baeldung.com/java-iterate-date-range)
 - [Initialize a HashMap in Java](https://www.baeldung.com/java-initialize-hashmap)
 - [Immutable Set in Java](https://www.baeldung.com/java-immutable-set)
-- [Filtering a Stream of Optionals in Java](https://www.baeldung.com/java-filter-stream-of-optional)
 
 Note: also contains part of the code for the article
 [How to Filter a Collection in Java](https://www.baeldung.com/java-collection-filtering).
