+++ core-java-modules/core-java-optional/README.md
@@ -1,8 +1,13 @@
-=========
+## Java Optional
 
-## Core Java Optional
+This module contains articles about Java Optional.
 
 ### Relevant Articles: 
 - [Java Optional as Return Type](https://www.baeldung.com/java-optional-return)
-- [Guide To Java 8 Optional](https://www.baeldung.com/java-optional)
+- [Guide to Java 8 Optional](https://www.baeldung.com/java-optional)
 - [Java Optional – orElse() vs orElseGet()](https://www.baeldung.com/java-optional-or-else-vs-or-else-get)
+- [Transforming an Empty String into an Empty Optional](https://www.baeldung.com/java-empty-string-to-empty-optional)
+- [Filtering a Stream of Optionals in Java](https://www.baeldung.com/java-filter-stream-of-optional)
+- [Java 9 Optional API Additions](https://www.baeldung.com/java-9-optional)
+- [Throw Exception in Optional in Java 8](https://www.baeldung.com/java-optional-throw-exception)
+- [Optional orElse Optional](https://www.baeldung.com/java-optional-or-else-optional)
