+++ core-java-modules/core-java-optional/src/main/java/com/baeldung/optional/orelse/OrElseAndOrElseGetBenchmarkRunner.java
@@ -1,4 +1,4 @@
-package com.baeldung.optional;
+package com.baeldung.optional.orelse;
 
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
