+++ core-java-modules/core-java-optional/src/main/java/com/baeldung/optionalreturntype/User.java
@@ -1,4 +1,4 @@
-package com.baeldung.optionalReturnType;
+package com.baeldung.optionalreturntype;
 
 import java.io.Serializable;
