+++ core-java-modules/core-java-optional/src/test/java/com/baeldung/filterstream/Java9OptionalsStreamUnitTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.java9;
+package com.baeldung.filterstream;
 
 import static org.junit.Assert.assertEquals;
 
@@ -47,15 +47,17 @@ public class Java9OptionalsStreamUnitTest {
         assertEquals("bar", filteredList.get(1));
     }
 
-    @Test
-    public void filterOutPresentOptionalsWithJava9() {
-        assertEquals(4, listOfOptionals.size());
 
-        List<String> filteredList = listOfOptionals.stream().flatMap(Optional::stream).collect(Collectors.toList());
-
-        assertEquals(2, filteredList.size());
-        assertEquals("foo", filteredList.get(0));
-        assertEquals("bar", filteredList.get(1));
-    }
+//    Uncomment code when code base is compatible with Java 9
+//    @Test
+//    public void filterOutPresentOptionalsWithJava9() {
+//        assertEquals(4, listOfOptionals.size());
+//
+//        List<String> filteredList = listOfOptionals.stream().flatMap(Optional::stream).collect(Collectors.toList());
+//
+//        assertEquals(2, filteredList.size());
+//        assertEquals("foo", filteredList.get(0));
+//        assertEquals("bar", filteredList.get(1));
+//    }
 
 }
