+++ core-java-modules/core-java-optional/src/test/java/com/baeldung/optional/OptionalChainingUnitTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.java8.optional;
+package com.baeldung.optional;
 
 import org.junit.Before;
 import org.junit.Test;
