+++ core-java-modules/core-java-optional/src/test/java/com/baeldung/optional/orelse/OrElseAndOrElseGetUnitTest.java
@@ -1,12 +1,10 @@
-package com.baeldung.java8.optional;
+package com.baeldung.optional.orelse;
 
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import static org.junit.Assert.*;
-
-import com.baeldung.optional.OrElseAndOrElseGet;
 
 public class OrElseAndOrElseGetUnitTest {
